+++ app/my/account/layout.tsx
@@ -1,8 +1,7 @@
 import type {Metadata} from "next";
-import ProtectRouter from "@/components/utils/ProtectRouter";
 
 export const metadata: Metadata = {
-  title: 'Job Seek | Sign Up'
+  title: 'Djinni | Account'
 }
 
 export default function AccountLayout({
@@ -11,8 +10,6 @@ export default function AccountLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <ProtectRouter allowedRoles={['Candidate']}>
-      <section>{children}</section>
-    </ProtectRouter>
+    <section>{children}</section>
   );
 }
