+++ app/my/account/page.tsx
@@ -1,28 +1,24 @@
 "use client";
 import {Separator} from "@/components/ui/separator"
-import {AccountForm} from "@/components/forms/account-form";
-import {useRetrieveMeCandidateQuery} from "@/lib/features/accounts/accountsApiSlice";
+import {useRetrieveUserQuery} from "@/lib/features/auth/authApiSlice";
 import Spinner from "@/components/general/Spinner";
-import {useListSkillsQuery} from "@/lib/features/other/otherApiSlice";
+import {AccountForm} from "@/components/forms/account-form";
 
 export default function Page() {
-  const {data: candidate, isLoading, isFetching} = useRetrieveMeCandidateQuery()
-  const {data: skills, isLoading: isLoadingSkills, isFetching: isFetchingSkills} = useListSkillsQuery();
-  if (isLoading || isFetching || isLoadingSkills || isFetchingSkills) {
-    return <Spinner size={150}/>
-  }
+  const {data: user, isLoading, isFetching} = useRetrieveUserQuery()
 
   return (
     <div className="space-y-6">
       <div>
-        <h3 className="text-lg font-medium">Contact and CV</h3>
+        <h3 className="text-lg font-medium">Account</h3>
         <p className="text-sm text-muted-foreground">
-          Job search on Djinni is anonymous.
-          Only those to whom you open contacts will see your personal data.
+          This is how others will see you on the site.
         </p>
       </div>
       <Separator/>
-      <AccountForm candidate={candidate} skills={skills}/>
+      {isLoading || isFetching ? <Spinner size={200}/> :
+        <AccountForm user={user}/>
+      }
     </div>
   )
 }
