+++ app/my/contacts/page.tsx
@@ -1,18 +1,18 @@
 import { Separator } from "@/components/ui/separator"
-import { ProfileForm } from "@/components/forms/profile-form"
+import WorkInProgress from "@/components/ui/work-in-progress";
 
 export default function Page() {
   return (
     <div className="space-y-6">
       <div>
-        <h3 className="text-lg font-medium">Contact and CV</h3>
+        <h3 className="text-lg font-medium">Contacts</h3>
         <p className="text-sm text-muted-foreground">
           Job search on Djinni is anonymous.
           Only those to whom you open contacts will see your personal data.
         </p>
       </div>
       <Separator />
-      <ProfileForm />
+      <WorkInProgress />
     </div>
   )
 }
