+++ app/my/profile/page.tsx
@@ -1,24 +1,31 @@
 "use client";
 import {Separator} from "@/components/ui/separator"
-import {ProfileForm} from "@/components/forms/profile-form"
-import {useRetrieveUserQuery} from "@/lib/features/auth/authApiSlice";
+import {ProfileForm} from "@/components/forms/profile-form";
+import {useRetrieveMeCandidateQuery} from "@/lib/features/accounts/accountsApiSlice";
 import Spinner from "@/components/general/Spinner";
+import {useListCategoryQuery, useListSkillsQuery} from "@/lib/features/other/otherApiSlice";
+import ImageForm from "@/components/forms/image-form";
 
 export default function Page() {
-  const {data: user, isLoading, isFetching} = useRetrieveUserQuery()
+  const {data: candidate, isLoading, isFetching} = useRetrieveMeCandidateQuery()
+  const {data: skills, isLoading: isLoadingSkills, isFetching: isFetchingSkills} = useListSkillsQuery();
+  const {data: category, isLoading: isLoadingCategory, isFetching: isFetchingCategory} = useListCategoryQuery();
+  if (isLoading || isFetching || isLoadingSkills || isFetchingSkills || isLoadingCategory || isFetchingCategory) {
+    return <Spinner size={150}/>
+  }
 
   return (
     <div className="space-y-6">
       <div>
         <h3 className="text-lg font-medium">Profile</h3>
         <p className="text-sm text-muted-foreground">
-          This is how others will see you on the site.
+          Job search on Djinni is anonymous.
+          Only those to whom you open contacts will see your personal data.
         </p>
       </div>
       <Separator/>
-      {isLoading || isFetching ? <Spinner size={200}/> :
-        <ProfileForm user={user}/>
-      }
+      <ImageForm candidate={candidate}/>
+      <ProfileForm candidate={candidate} skills={skills} category={category}/>
     </div>
   )
 }
